Move isInvalidShape method to annotation/utils.js
