+++ resources/assets/js/videos/utils.js
@@ -24,32 +24,4 @@ let getRoundToPrecision = function (reference) {
     };
 };
 
-/**
- * This function checks for invalid annotation shapes.
- * 
- * @param feature containing the video annotation to check
- * @returns true, if a video annotation has an invalid shape, otherwise false.
- * 
- * **/
-let isInvalidShape = function (feature) {
-    let geometry = feature.getGeometry();
-    let points = [];
-    switch (geometry.getType()) {
-        case 'Circle':
-            return parseInt(geometry.getRadius()) === 0;
-        case 'LineString':
-            points = geometry.getCoordinates();
-            return (new Set(points.map(xy => String([xy])))).size < 2;
-        case 'Rectangle':
-        case 'Ellipse':
-            points = geometry.getCoordinates()[0];
-            return (new Set(points.map(xy => String([xy])))).size !== 4;
-        case 'Polygon':
-            points = geometry.getCoordinates()[0];
-            return (new Set(points.map(xy => String([xy])))).size < 3;
-        default:
-            return false;
-    }
-};
-
-export {getRoundToPrecision, isInvalidShape};
+export {getRoundToPrecision};
